modifiying the logic

+++ src/main/java/com/api/zorvanz/domain/cart/CartData.java
@@ -2,12 +2,11 @@ package com.api.zorvanz.domain.cart;
 
 import com.api.zorvanz.domain.cartitem.CartItemData;
 
-import java.math.BigDecimal;
 import java.util.List;
 
 public record CartData(
         Long id,
-        BigDecimal totalAmount,
+        Double totalAmount,
         List <CartItemData> cartItems
 ) {
 }
